--- fourtoon-cookie/src/pages/IntroPage/AppleSignInAndSignUpButton/AppleSignInAndSignUpButton.styled.ts
@@ -1,26 +1,0 @@
-import { StyleSheet } from "react-native";
-
-
-export const styles = StyleSheet.create({
-    appleButton: {
-        flexDirection: 'row',
-        justifyContent: 'center',
-        alignItems: 'center',
-        padding: 19,
-        backgroundColor: '#000000',
-        borderRadius: 16,
-      },
-      appleLogo: {
-        width: 24,
-        height: 24,
-        position: 'absolute',
-        left: 18,
-      },
-      appleButtonText: {
-        fontFamily: 'Pretendard',
-        fontStyle: 'normal',
-        fontWeight: '600',
-        fontSize: 17,
-        color: '#FFFFFF',
-      },
-})
